refactor: Message Input 커스텀 컴포넌트 & hooks 파일 생성
