--- src/components/Common/MessageImage/MessageImage.style.ts
@@ -1,12 +1,0 @@
-import styled from '@emotion/styled';
-
-export const MessageImage = styled.img`
-    display: inline-block;
-    margin: 5px;
-    border-radius: 10px;
-    height: 60px;
-
-    &:hover {
-        cursor: pointer;
-    }
-`
