--- src/components/Common/MessageImage/MessageImage.tsx
@@ -1,5 +1,0 @@
-import { MessageImage } from '@/components/Common/MessageImage/MessageImage.style.ts';
-
-export default function MessageImageStyle({ src, alt, onClick }) {
-  return <MessageImage src={src} alt={alt} onClick={onClick} />
-}
