+++ src/components/Order/Message/Message.style.ts
@@ -40,3 +40,14 @@ export const ErrorMessage = styled.p`
     padding-bottom: 30px;
     color: ${({ theme }) => theme.colors.red700};
 `
+
+export const MessageImage = styled.img`
+    display: inline-block;
+    margin: 5px;
+    border-radius: 10px;
+    height: 60px;
+
+    &:hover {
+        cursor: pointer;
+    }
+`
